requirements.txt
streamlit
opencv-python
numpy
pandas
scikit-image
scikit-learn
Pillow
matplotlib
seaborn
opencv-python-headless